clean up mock weights
