+++ frame/elections-phragmen/src/weights.rs
@@ -229,85 +229,3 @@ impl WeightInfo for () {
 	}
 }
 
-pub struct MockWeightInfo<T>(PhantomData<T>);
-impl<T: frame_system::Config> WeightInfo for MockWeightInfo<T> {
-	fn vote_equal(v: u32, ) -> Weight {
-		(45_157_000 as Weight)
-			// Standard Error: 6_000
-			.saturating_add((399_000 as Weight).saturating_mul(v as Weight))
-			.saturating_add(RocksDbWeight::get().reads(5 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(2 as Weight))
-	}
-	fn vote_more(v: u32, ) -> Weight {
-		(69_738_000 as Weight)
-			// Standard Error: 14_000
-			.saturating_add((450_000 as Weight).saturating_mul(v as Weight))
-			.saturating_add(RocksDbWeight::get().reads(5 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(2 as Weight))
-	}
-	fn vote_less(v: u32, ) -> Weight {
-		(73_955_000 as Weight)
-			// Standard Error: 38_000
-			.saturating_add((227_000 as Weight).saturating_mul(v as Weight))
-			.saturating_add(RocksDbWeight::get().reads(5 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(2 as Weight))
-	}
-	fn remove_voter() -> Weight {
-		(68_398_000 as Weight)
-			.saturating_add(RocksDbWeight::get().reads(2 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(2 as Weight))
-	}
-	fn submit_candidacy(c: u32, ) -> Weight {
-		(59_291_000 as Weight)
-			// Standard Error: 2_000
-			.saturating_add((412_000 as Weight).saturating_mul(c as Weight))
-			.saturating_add(RocksDbWeight::get().reads(3 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(1 as Weight))
-	}
-	fn renounce_candidacy_candidate(c: u32, ) -> Weight {
-		(55_026_000 as Weight)
-			// Standard Error: 2_000
-			.saturating_add((207_000 as Weight).saturating_mul(c as Weight))
-			.saturating_add(RocksDbWeight::get().reads(1 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(1 as Weight))
-	}
-	fn renounce_candidacy_members() -> Weight {
-		(77_840_000 as Weight)
-			.saturating_add(RocksDbWeight::get().reads(4 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(4 as Weight))
-	}
-	fn renounce_candidacy_runners_up() -> Weight {
-		(54_559_000 as Weight)
-			.saturating_add(RocksDbWeight::get().reads(1 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(1 as Weight))
-	}
-	fn remove_member_with_replacement() -> Weight {
-		(84_311_000 as Weight)
-			.saturating_add(RocksDbWeight::get().reads(5 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(5 as Weight))
-	}
-	fn remove_member_wrong_refund() -> Weight {
-		(7_677_000 as Weight)
-			.saturating_add(RocksDbWeight::get().reads(1 as Weight))
-	}
-	fn clean_defunct_voters(v: u32, d: u32, ) -> Weight {
-		(0 as Weight)
-			// Standard Error: 55_000
-			.saturating_add((114_815_000 as Weight).saturating_mul(v as Weight))
-			// Standard Error: 53_000
-			.saturating_add((49_000 as Weight).saturating_mul(d as Weight))
-			.saturating_add(RocksDbWeight::get().reads(4 as Weight))
-			.saturating_add(RocksDbWeight::get().reads((3 as Weight).saturating_mul(v as Weight)))
-			.saturating_add(RocksDbWeight::get().writes((3 as Weight).saturating_mul(v as Weight)))
-	}
-	fn election_phragmen(c: u32, v: u32, e: u32, ) -> Weight {
-		(0 as Weight)
-			// Standard Error: 1_940_000
-			.saturating_add((1 as Weight).saturating_mul(c as Weight))
-			// Standard Error: 807_000
-			.saturating_add((1 as Weight).saturating_mul(v as Weight))
-			// Standard Error: 55_000
-			.saturating_add((1 as Weight).saturating_mul(e as Weight))
-	}
-}
-
